requirements.txt
# ───── LLM Providers & Environment ─────
python-dotenv>=1.0.0     # Robust environment variable management
openai>=1.0.0           # OpenAI official SDK
google-generativeai>=0.3.0  # Google AI official SDK for Gemini

# ───── OCR y Preprocesamiento ─────
fasttext-wheel>=0.9.2    # Reemplazo más robusto para langdetect
langdetect==1.0.9       # Mantenido temporalmente para compatibilidad
opencv-python>=4.8.1.78  # Versión estable con parches de seguridad
Pillow>=10.0.0          # Versión con correcciones de seguridad
pytesseract>=0.3.10
unidecode>=1.3.8
numpy>=1.24.3

# ───── PDF Parsing ─────
camelot-py>=0.11.0      # Mantener por su robustez en extracción de tablas
pdfplumber>=0.10.3      # Para análisis detallado de PDFs
PyMuPDF>=1.23.0         # Última versión estable con mejoras de rendimiento
pypdf>=3.16.0           # Versión actualizada sin warnings de ARC4
cryptography>=42.0.0    # Versión futura que elimina ARC4 deprecado
charset_normalizer>=3.3.2  # Requerido por pdfminer

# ───── NLP y Detección de Errores OCR ─────
nltk>=3.8.1             # Versión con parches de seguridad
symspellpy>=6.7.0       # Reemplazo más eficiente para pyspellchecker
spacy>=3.7.4            # Mantener spaCy como núcleo NLP
fasttext-wheel>=0.9.2   # Detector de idiomas mejorado

# ───── Machine Learning ─────
matplotlib>=3.9.0
plotly>=5.18.0          # Para visualizaciones interactivas

# ───── LLM Integration ─────
openai>=1.12.0          # SDK oficial de OpenAI
google-generativeai>=0.3.0  # SDK oficial de Google para Gemini
langchain>=0.1.0        # Para mejor gestión de prompts y LLMs
pydantic-settings>=2.0.0  # Para validación de configuraciones
scikit-learn>=1.5.0     # Base para ML clásico

# ───── Logging y Utilidades ─────
loguru==0.7.3
tabulate==0.9.0
python-dotenv==1.0.0
markdown2==2.4.10
jinja2==3.1.3

# ───── Testing ─────
fasttext-wheel==0.9.2
pandas==2.2.2
pytest==7.2.0